Add new menu system; icons
