+++ src/app/mosaic/page.module.css
@@ -8,6 +8,7 @@
   text-align: center;
   margin-bottom: var(--space-12);
   color: var(--gray-700);
+  padding-top: var(--space-4);
 
   h1 {
     font-size: var(--text-4xl);
@@ -77,6 +78,9 @@
   height: auto;
   margin: 0 auto;
   width: fit-content;
+  grid-gap: 0;
+  column-gap: 0;
+  row-gap: 0;
 }
 
 .mosaicItem {
@@ -92,8 +96,6 @@
   display: block;
   width: 100%;
   height: 100%;
-  max-width: 200px;
-  max-height: 200px;
 
   &:hover {
     transform: scale(1.05);
