+++ src/app/page.tsx
@@ -1,23 +1,12 @@
-"use client";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-import AuthLoading from "src/components/AuthLoading/AuthLoading.component";
-import Login from "src/components/Login/Login.component";
-import { useAuth } from "src/context/auth.context";
+import type { Metadata } from "next";
+import HomeClient from "src/components/HomeClient/HomeClient.component";
+
+export const metadata: Metadata = {
+  title: "FilterMyDisco.gs - A Discogs collection management tool",
+  description:
+    "View, filter and sort your Discogs collection and build a crate as you browse",
+};
 
 export default function Home() {
-  const { state } = useAuth();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (state.isAuthenticated) {
-      router.replace("/releases");
-    }
-  }, [state.isAuthenticated, router]);
-
-  if (state.isLoading) {
-    return <AuthLoading />;
-  }
-
-  return <Login />;
+  return <HomeClient />;
 }
