+++ src/components/AuthLoading/AuthLoading.component.tsx
@@ -1,11 +1,12 @@
+import Logo from "src/styles/icons/fmd-logo.svg";
 import styles from "./AuthLoading.module.css";
 
 export const AuthLoading = () => {
   return (
     <div className={styles.container}>
       <div className={styles.loadingCard}>
+        <Logo className={styles.logo} />
         <div className={styles.spinner} />
-        <h1 className={styles.title}>Filter My Disco.gs</h1>
         <p className={styles.description}>Checking authentication...</p>
       </div>
     </div>
