+++ src/components/MosaicClient/MosaicClient.module.css
@@ -1,0 +1,246 @@
+.container {
+  min-height: 100vh;
+  padding: var(--space-8);
+  background: var(--white);
+}
+
+.header {
+  text-align: center;
+  margin-bottom: var(--space-12);
+  color: var(--gray-700);
+  padding-top: var(--space-4);
+
+  h1 {
+    font-size: var(--text-4xl);
+    font-weight: var(--font-bold);
+    margin-bottom: var(--space-4);
+    color: var(--black);
+
+    @media (min-width: 768px) {
+      font-size: 3rem;
+    }
+  }
+
+  p {
+    font-size: var(--text-lg);
+    margin-bottom: var(--space-8);
+    color: var(--gray-600);
+
+    @media (min-width: 768px) {
+      font-size: 1.2rem;
+    }
+  }
+}
+
+.downloadButton {
+  background: var(--primary-600);
+  border: none;
+  border-radius: var(--radius-lg);
+  color: var(--white);
+  cursor: pointer;
+  font-size: var(--text-lg);
+  font-weight: var(--font-semibold);
+  padding: var(--space-4) var(--space-8);
+  transition: var(--transition-normal);
+  box-shadow: var(--shadow-md);
+
+  &:hover:not(:disabled) {
+    background: var(--primary-700);
+    transform: translateY(-1px);
+    box-shadow: var(--shadow-lg);
+  }
+
+  &:disabled {
+    opacity: 0.7;
+    cursor: not-allowed;
+    transform: none;
+  }
+}
+
+.generatingContent {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: var(--space-2);
+  width: 100%;
+}
+
+.generatingText {
+  font-size: var(--text-base);
+  font-weight: var(--font-medium);
+}
+
+.progressBar {
+  width: 100%;
+  height: 4px;
+  background: rgba(255, 255, 255, 0.3);
+  border-radius: 2px;
+  overflow: hidden;
+}
+
+.progressFill {
+  height: 100%;
+  background: var(--white);
+  border-radius: 2px;
+  transition: width 0.3s ease;
+}
+
+.mosaicContainer {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  overflow: auto;
+  min-height: 400px;
+  padding: var(--space-2);
+
+  @media (min-width: 1000px) {
+    max-width: 95vw;
+    margin: 0 auto;
+  }
+}
+
+.mosaicGrid {
+  display: grid;
+  gap: 0;
+  height: auto;
+  margin: 0 auto;
+  width: fit-content;
+  grid-gap: 0;
+  column-gap: 0;
+  row-gap: 0;
+}
+
+.mosaicItem {
+  aspect-ratio: 1;
+  overflow: hidden;
+  transition: var(--transition-fast);
+  cursor: pointer;
+  position: relative;
+  background: none;
+  border: none;
+  padding: 0;
+  margin: 0;
+  display: block;
+  width: 100%;
+  height: 100%;
+
+  &:hover {
+    transform: scale(1.05);
+    z-index: 10;
+    position: relative;
+  }
+}
+
+.mosaicImage {
+  height: 100%;
+  object-fit: cover;
+  transition: var(--transition-fast);
+  width: 100%;
+  display: block;
+
+  .mosaicItem:hover & {
+    filter: brightness(1.1);
+  }
+}
+
+.loadingOverlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: var(--radius-sm);
+
+  .loadingSpinner {
+    width: 20px;
+    height: 20px;
+    border: 2px solid var(--gray-200);
+    border-top: 2px solid var(--primary-600);
+    border-radius: 50%;
+    animation: spin 1s linear infinite;
+    margin: 0;
+  }
+}
+
+.loadingContainer {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  justify-content: center;
+  text-align: center;
+}
+
+.mainLoadingSpinner {
+  animation: spin 1s linear infinite;
+  border: 4px solid var(--gray-200);
+  border-radius: 50%;
+  border-top: 4px solid var(--primary-600);
+  height: 50px;
+  margin-bottom: var(--space-4);
+  width: 50px;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+.emptyState {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  justify-content: center;
+  text-align: center;
+
+  h1 {
+    color: var(--black);
+    font-size: var(--text-4xl);
+    font-weight: var(--font-bold);
+    margin-bottom: var(--space-4);
+  }
+
+  p {
+    color: var(--gray-600);
+    font-size: var(--text-lg);
+  }
+}
+
+/* Mobile-first responsive design */
+@media (max-width: 768px) {
+  .container {
+    padding: var(--space-4);
+  }
+
+  .header {
+    h1 {
+      font-size: var(--text-3xl);
+    }
+
+    p {
+      font-size: var(--text-base);
+    }
+  }
+
+  .mosaicContainer {
+    padding: var(--space-4);
+    overflow-x: auto;
+    width: 100%;
+  }
+}
+
+@media (max-width: 480px) {
+  .mosaicContainer {
+    padding: var(--space-2);
+  }
+}
