+++ src/components/ReleasesClient/ReleasesClient.module.css
@@ -1,0 +1,146 @@
+.container {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-template-areas: "main";
+  width: 100%;
+  min-height: calc(100vh - 60px); /* Subtract header height */
+
+  &.withSidebar {
+    grid-template-columns: 1fr 400px;
+    grid-template-areas: "main sidebar";
+
+    @media (max-width: 480px) {
+      grid-template-columns: 1fr;
+      grid-template-areas: "main";
+    }
+  }
+}
+
+.mainContent {
+  background-color: var(--white);
+  grid-area: main;
+  overflow-y: auto;
+  padding: 0 var(--space-4);
+}
+
+.sidebar {
+  grid-area: sidebar;
+  display: none;
+
+  .withSidebar & {
+    display: block;
+  }
+
+  @media (max-width: 1023px) {
+    display: none;
+  }
+}
+
+.releasesHeader {
+  align-items: center;
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: var(--space-4);
+  width: 100%;
+  padding-top: var(--space-4);
+}
+
+.releasesHeader p {
+  align-items: center;
+  color: var(--gray-500);
+  display: flex;
+  font-size: var(--text-base);
+  gap: var(--space-2);
+  margin: 0;
+}
+
+.loadingIcon {
+  display: inline-flex;
+  gap: var(--space-2);
+  align-items: center;
+}
+
+.spinner {
+  animation: spin 1s linear infinite;
+  border: 2px solid var(--gray-100);
+  border-radius: var(--radius-full);
+  border-top-color: var(--gray-500);
+  height: 16px;
+  width: 16px;
+}
+
+@keyframes spin {
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+.releasesGrid {
+  align-items: stretch;
+  display: grid;
+  gap: var(--space-4);
+  grid-template-columns: 1fr;
+  justify-content: stretch;
+  width: 100%;
+  content-visibility: visible;
+  contain-intrinsic-size: 400px 300px;
+
+  & > div {
+    scroll-margin-top: 150px;
+  }
+
+  @media (min-width: 620px) {
+    gap: var(--space-4);
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  }
+}
+
+.releasesList {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-3);
+  width: 100%;
+  content-visibility: auto;
+  contain-intrinsic-size: 1px 800px;
+
+  & > div {
+    scroll-margin-top: 150px;
+  }
+}
+
+.errorContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 50vh;
+  padding: var(--space-8);
+  text-align: center;
+}
+
+.errorContainer h2 {
+  color: var(--error-500);
+  margin-bottom: var(--space-4);
+}
+
+.emptyState {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 50vh;
+  padding: var(--space-8);
+  text-align: center;
+}
+
+.emptyState h2 {
+  color: var(--gray-500);
+  margin-bottom: var(--space-4);
+}
+
+.loadingTrigger {
+  display: flex;
+  justify-content: center;
+  padding: var(--space-8);
+  min-height: 100px;
+}
