+++ src/components/StickyHeaderBar/components/FiltersBar.module.css
@@ -1,0 +1,48 @@
+.filtersBar {
+  background: var(--gray-50);
+  border-bottom: 1px solid var(--gray-200);
+  padding: var(--space-2) 0;
+  position: sticky;
+  top: 82px;
+  width: 100%;
+  z-index: 100;
+
+  .desktopFilters {
+    display: none;
+
+    @media (min-width: 769px) {
+      align-items: center;
+      display: flex;
+      flex-direction: row;
+      gap: var(--space-4);
+      justify-content: center;
+      margin: 0 auto;
+      max-width: 1200px;
+      padding: 0 var(--space-4);
+    }
+  }
+
+  .mobileFilters {
+    display: none;
+  }
+
+  .crateCount {
+    align-items: center;
+    background: var(--white);
+    border-radius: 50%;
+    color: var(--primary-600);
+    display: flex;
+    font-size: 10px;
+    font-weight: var(--font-bold);
+    height: 18px;
+    justify-content: center;
+    min-width: 18px;
+    padding: 0;
+
+    @media (min-width: 769px) {
+      font-size: var(--text-xs);
+      height: 20px;
+      min-width: 20px;
+    }
+  }
+}
