+++ src/components/StickyHeaderBar/components/FiltersBar.tsx
@@ -1,0 +1,154 @@
+import { useState } from "react";
+import { trackEvent } from "src/analytics/analytics";
+import AutocompleteSelect from "src/components/AutocompleteSelect/AutocompleteSelect.component";
+import Button from "src/components/Button/Button.component";
+import FiltersDrawer from "src/components/FiltersDrawer/FiltersDrawer.component";
+import Select from "src/components/Select/Select.component";
+import { useCollectionContext } from "src/context/collection.context";
+import { useCrate } from "src/context/crate.context";
+import { useFilterHandlers } from "src/hooks/useFilterHandlers.hook";
+import styles from "./FiltersBar.module.css";
+
+interface FiltersBarProps {
+  category: string;
+  disabled?: boolean;
+  showCrate?: boolean;
+}
+
+export const FiltersBar = ({
+  category,
+  disabled = false,
+  showCrate = true,
+}: FiltersBarProps) => {
+  const { state: collectionState } = useCollectionContext();
+  const { selectedReleases, openDrawer } = useCrate();
+  const [isFiltersDrawerOpen, setIsFiltersDrawerOpen] = useState(false);
+
+  const { fetchingCollection, collection, error } = collectionState;
+
+  const {
+    handleStyleChange,
+    handleYearChange,
+    handleSortChange,
+    styleOptions,
+    yearOptions,
+    selectedStyles,
+    selectedYears,
+    selectedSort,
+  } = useFilterHandlers(category);
+
+  const handleFiltersClick = () => {
+    setIsFiltersDrawerOpen(true);
+    trackEvent("filtersOpened", {
+      action: "filtersOpenedFromHeader",
+      category: "mobile_filters",
+      label: "Filters Opened from Header",
+      value: "mobile",
+    });
+  };
+
+  const closeFiltersDrawer = () => {
+    setIsFiltersDrawerOpen(false);
+  };
+
+  const handleCrateClick = () => {
+    openDrawer();
+    trackEvent("crateOpened", {
+      action: "crateOpenedFromHeader",
+      category: "crate",
+      label: "Crate Opened from Header",
+      value: selectedReleases.length.toString(),
+    });
+  };
+
+  const isDisabled = disabled || fetchingCollection || !collection || error;
+
+  if (styleOptions.length === 0 || isDisabled) {
+    return null;
+  }
+
+  return (
+    <>
+      <div className={styles.filtersBar}>
+        {/* Desktop filters */}
+        <div className={styles.desktopFilters}>
+          <AutocompleteSelect
+            label="Style"
+            options={styleOptions}
+            value={selectedStyles}
+            onChange={handleStyleChange}
+            disabled={!collection}
+            multiple={true}
+            placeholder="Select styles..."
+          />
+          <AutocompleteSelect
+            label="Year"
+            options={yearOptions}
+            value={selectedYears.map((year) => year.toString())}
+            onChange={handleYearChange}
+            disabled={!collection}
+            multiple={true}
+            placeholder="All years"
+          />
+          <Select
+            label="Sort by"
+            options={[]} // Will be populated by parent
+            value={selectedSort}
+            onChange={handleSortChange}
+            disabled={fetchingCollection}
+            placeholder="Select sort option..."
+          />
+
+          {/* Crate button in desktop filters */}
+          {showCrate && selectedReleases.length > 0 && (
+            <Button
+              variant="primary"
+              size="md"
+              onPress={handleCrateClick}
+              aria-label={`Open crate with ${selectedReleases.length} items`}
+            >
+              <span>My Crate</span>
+              <span className={styles.crateCount}>
+                {selectedReleases.length}
+              </span>
+            </Button>
+          )}
+        </div>
+
+        {/* Mobile filters button */}
+        <div className={styles.mobileFilters}>
+          <Button
+            variant="secondary"
+            size="sm"
+            onPress={handleFiltersClick}
+            aria-label="Open filters"
+          >
+            <span>⚙️</span>
+            <span>Filters</span>
+          </Button>
+
+          {/* Crate button in mobile filters */}
+          {showCrate && selectedReleases.length > 0 && (
+            <Button
+              variant="primary"
+              size="sm"
+              onPress={handleCrateClick}
+              aria-label={`Open crate with ${selectedReleases.length} items`}
+            >
+              <span>My Crate</span>
+              <span className={styles.crateCount}>
+                {selectedReleases.length}
+              </span>
+            </Button>
+          )}
+        </div>
+      </div>
+
+      {/* Mobile filters drawer */}
+      <FiltersDrawer
+        isOpen={isFiltersDrawerOpen}
+        onClose={closeFiltersDrawer}
+      />
+    </>
+  );
+};
