+++ src/components/StickyHeaderBar/components/MobileFiltersButton.module.css
@@ -1,0 +1,1 @@
+/* This component uses the mobile actions styles from UserActions */
